requirements.txt
Flask
Flask-PyMongo
PyJWT
bcrypt
python-dotenv
Flask-Limiter
Flask-Cors